+++ packages/markets/src/components/favoritesListFull/favoritesListFull.ui.tsx
@@ -1,11 +1,11 @@
 import { FC } from "react";
-import { cn, Flex, DataTable } from "@orderly.network/ui";
-import { UnFavoritesIcon } from "../../icons";
-import { UseFavoritesListFullReturn } from "./favoritesListFull.script";
-import { useMarketsContext } from "../../components/marketsProvider";
-import { FavoritesTabWidget } from "../../components/favoritesTabs";
-import { useFavoritesListFullColumns } from "./column";
 import { Trans } from "@orderly.network/i18n";
+import { cn, Flex, DataTable } from "@orderly.network/ui";
+import { FavoritesTabWidget } from "../../components/favoritesTabs";
+import { useMarketsContext } from "../../components/marketsProvider";
+import { UnFavoritesIcon } from "../../icons";
+import { useFavoritesListFullColumns } from "./column";
+import { UseFavoritesListFullReturn } from "./favoritesListFull.script";
 
 export type FavoritesListFullProps = UseFavoritesListFullReturn;
 
@@ -21,7 +21,9 @@ export const FavoritesListFull: FC<FavoritesListFullProps> = (props) => {
       {/* @ts-ignore */}
       <Trans
         i18nKey="markets.dataList.favorites.empty"
-        components={[<UnFavoritesIcon className="oui-text-base-contrast-36" />]}
+        components={[
+          <UnFavoritesIcon key="0" className="oui-text-base-contrast-36" />,
+        ]}
       />
     </Flex>
   );