+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/core
 
+## 2.1.3
+
+### Patch Changes
+
+- publish
+- 5ba2a31a8: publish
+- Updated dependencies
+- Updated dependencies [5ba2a31a8]
+  - @orderly.network/types@2.1.3
+
+## 2.1.3-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.1.3-alpha.0
+
 ## 2.1.2
 
 ### Patch Changes